refactoring + working on chainstate trx + format
