+++ ab-tester/justfile
@@ -2,3 +2,6 @@ set dotenv-load
 
 run *args:
     RUST_BACKTRACE=1 RUST_LOG="debug" cargo run -- {{args}}
+
+fmt:
+    cargo fmt --all -- --config group_imports=StdExternalCrate,imports_granularity=Module
