+++ ab-tester/src/cli.rs
@@ -8,7 +8,8 @@ use clap_verbosity_flag::Verbosity;
 use color_eyre::eyre::{bail, Result};
 
 use crate::context::callbacks::DefaultReplayCallbacks;
-use crate::context::{replay::ReplayOpts, Runtime};
+use crate::context::replay::ReplayOpts;
+use crate::context::Runtime;
 use crate::ok;
 
 /// Our CLI entrypoint.
