+++ ab-tester/src/cli/console/screens.rs
@@ -2,7 +2,8 @@ pub mod blocks;
 pub mod main;
 pub mod start;
 
-pub use {blocks::BlocksScreen, start::StartScreen};
+pub use blocks::BlocksScreen;
+pub use start::StartScreen;
 
 pub enum Screen {
     Start,
