+++ ab-tester/src/cli/console/screens/main.rs
@@ -1,4 +1,5 @@
-use std::{cell::RefCell, rc::Rc};
+use std::cell::RefCell;
+use std::rc::Rc;
 
 use ratatui::prelude::{Constraint, Layout};
