+++ ab-tester/src/context/callbacks.rs
@@ -1,4 +1,5 @@
-use super::{environments::RuntimeEnv, Network};
+use super::environments::RuntimeEnv;
+use super::Network;
 
 #[allow(unused_variables)]
 pub trait RuntimeEnvCallbackHandler {
