+++ ab-tester/src/context/environments/instrumented.rs
@@ -1,34 +1,32 @@
-use std::{cell::RefCell, rc::Rc};
+use std::cell::RefCell;
+use std::rc::Rc;
 
-use color_eyre::{
-    eyre::{anyhow, bail},
-    Result,
-};
+use color_eyre::eyre::{anyhow, bail};
+use color_eyre::Result;
 use diesel::{
     Connection, ExpressionMethods, OptionalExtension, QueryDsl, RunQueryDsl, SqliteConnection,
 };
 use log::*;
 
-use crate::{
-    context::{
-        blocks::BlockHeader,
-        boot_data::mainnet_boot_data,
-        callbacks::{DefaultEnvCallbacks, RuntimeEnvCallbackHandler},
-        environments::BlockTransactionContext,
-        Block, BlockCursor, Network, Runtime, TestEnvPaths,
-    },
-    db::model::app_db as model,
-    db::{appdb::AppDb, schema::appdb},
-    ok, stacks,
-};
-
 use super::{ReadableEnv, RuntimeEnv, WriteableEnv};
+use crate::context::blocks::BlockHeader;
+use crate::context::boot_data::mainnet_boot_data;
+use crate::context::callbacks::{DefaultEnvCallbacks, RuntimeEnvCallbackHandler};
+use crate::context::{
+    Block, BlockCursor, BlockTransactionContext, Network, RegularBlockTransactionContext, Runtime,
+    StacksEnvPaths,
+};
+use crate::db::appdb::AppDb;
+use crate::db::model::app_db as model;
+use crate::db::schema::appdb;
+use crate::db::stacks_node::StacksNodeDb;
+use crate::{clarity, ok, stacks};
 
 /// Holds the configuration of an [InstrumentedEnv].
 pub struct InstrumentedEnvConfig {
     working_dir: String,
     readonly: bool,
-    paths: TestEnvPaths,
+    paths: StacksEnvPaths,
     runtime: Runtime,
     network: Network,
 }
@@ -39,6 +37,7 @@ pub struct InstrumentedEnvState {
     chainstate: stacks::StacksChainState,
     clarity_db_conn: SqliteConnection,
     sortition_db: stacks::SortitionDB,
+    burnstate_db: Box<dyn clarity::BurnStateDB>,
 }
 
 /// This environment type is app-specific and will instrument all Clarity-related
@@ -64,7 +63,7 @@ impl InstrumentedEnv {
         runtime: Runtime,
         network: Network,
     ) -> Result<Self> {
-        let paths = TestEnvPaths::new(&working_dir);
+        let paths = StacksEnvPaths::new(&working_dir);
 
         let env_config = InstrumentedEnvConfig {
             working_dir,
@@ -176,6 +175,7 @@ impl InstrumentedEnv {
     }
 }
 
+/// Implementation of [RuntimeEnv] for [InstrumentedEnv].
 impl RuntimeEnv for InstrumentedEnv {
     fn name(&self) -> String {
         self.name.clone()
@@ -211,11 +211,11 @@ impl RuntimeEnv for InstrumentedEnv {
 
         debug!("initializing chainstate");
         self.callbacks
-            .open_chainstate_start(self, &paths.chainstate_path);
+            .open_chainstate_start(self, &paths.chainstate_dir);
         let (chainstate, _) = stacks::StacksChainState::open_and_exec(
             network.is_mainnet(),
             1,
-            &paths.chainstate_path,
+            &paths.chainstate_dir,
             Some(&mut boot_data),
             Some(marf_opts.clone()),
         )?;
@@ -239,16 +239,22 @@ impl RuntimeEnv for InstrumentedEnv {
         //debug!("attempting to migrate sortition db");
         debug!("opening sortition db");
         self.callbacks
-            .open_sortition_db_start(self, &paths.sortition_db_path);
-        let sortition_db = super::open_sortition_db(&paths.sortition_db_path, network)?;
+            .open_sortition_db_start(self, &paths.sortition_dir);
+        let sortition_db = super::open_sortition_db(&paths.sortition_dir, network)?;
         self.callbacks.open_sortition_db_finish(self);
         info!("successfully opened sortition db");
+
+        let burnstate_db: Box<dyn clarity::BurnStateDB> = Box::new(StacksNodeDb::new(
+            &paths.index_db_path,
+            &paths.sortition_db_path,
+        )?);
 
         let state = InstrumentedEnvState {
             chainstate,
             index_db_conn: RefCell::new(index_db_conn),
             clarity_db_conn,
             sortition_db,
+            burnstate_db,
         };
 
         self.env_state = Some(state);
@@ -268,11 +274,7 @@ impl ReadableEnv for InstrumentedEnv {
 
 /// Implementation of [WriteableEnv] for [InstrumentedEnv].
 impl WriteableEnv for InstrumentedEnv {
-    fn block_begin(
-        &mut self,
-        block: &crate::context::Block,
-        f: impl FnOnce(&mut super::BlockTransactionContext) -> Result<()>,
-    ) -> Result<()> {
+    fn block_begin(&mut self, block: &crate::context::Block) -> Result<BlockTransactionContext> {
         if self.is_readonly() {
             bail!("[{}] environment is read-only.", self.name);
         }
@@ -306,44 +308,58 @@ impl WriteableEnv for InstrumentedEnv {
         // initialization. Log that we reached it and skip processing.
         if let Block::Genesis(_) = block {
             info!("genesis block cannot be processed as it was statically initialized; moving on");
-            return ok!();
+            return Ok(BlockTransactionContext::Genesis);
         }
 
         let state = self.get_env_state_mut()?;
 
         // Get an instance to the BurnStateDB (SortitionDB's `index_conn` implements this trait).
-        let burn_db = state.sortition_db.index_conn();
 
-        // Start a new chainstate transaction.
-        debug!("creating chainstate tx");
-        let (chainstate_tx, clarity_instance) = state.chainstate.chainstate_tx_begin()?;
-        debug!("chainstate tx started");
-
-        // Begin a new Clarity block.
-        debug!("beginning clarity block");
-        let mut clarity_block_conn = clarity_instance.begin_block(
+        let mut clarity_block_conn = state.chainstate.clarity_state.begin_block(
             &current_block_id,
             &next_block_id,
-            &chainstate_tx,
-            &burn_db,
+            &state.chainstate.state_index,
+            &*state.burnstate_db,
         );
+
+        // Start a new chainstate transaction.
+        //debug!("creating chainstate tx");
+        //let (chainstate_tx, clarity_instance) = state.chainstate.chainstate_tx_begin()?;
+        //debug!("chainstate tx started");
+
+        // Begin a new Clarity block.
+        //debug!("beginning clarity block");
+        /*let mut clarity_block_conn = clarity_instance.begin_block(
+            &current_block_id,
+            &next_block_id,
+            &state.chainstate.state_index,
+            &*state.burnstate_db,
+        );*/
 
         // Enter Clarity transaction processing for the new block.
         debug!("starting clarity tx processing");
-        let clarity_tx_conn = clarity_block_conn.start_transaction_processing();
+        //let clarity_tx_conn = ;
 
         // Call the provided function with our context object.
         //let mut block_tx_ctx = ;
-        f(&mut BlockTransactionContext {
-            clarity_tx_conn: &clarity_tx_conn,
-        })?;
+        let block_tx_ctx =
+            RegularBlockTransactionContext::new(block.stacks_block_id()?, clarity_block_conn);
 
-        debug!("returning");
+        Ok(BlockTransactionContext::Regular(block_tx_ctx))
+    }
 
-        clarity_tx_conn.commit();
-        clarity_block_conn.commit_to_block(&block.stacks_block_id()?);
-        chainstate_tx.commit()?;
+    fn block_commit(
+        &mut self,
+        block_tx_ctx: BlockTransactionContext,
+    ) -> Result<clarity::LimitedCostTracker> {
+        if let BlockTransactionContext::Regular(ctx) = block_tx_ctx {
+            //clarity_tx_conn.commit();
+            //let cost_tracker = clarity_block_conn.commit_to_block(&stacks_block_id);
+            //chainstate_tx.commit()?;
+        } else {
+            bail!("Cannot commit genesis block as it has already been statically loaded from boot data.")
+        }
 
-        ok!()
+        todo!()
     }
 }
