+++ ab-tester/src/db.rs
@@ -2,3 +2,4 @@ pub mod appdb;
 pub mod datastore;
 pub mod model;
 pub mod schema;
+pub mod stacks_node;
