+++ ab-tester/src/db/schema/clarity_marf.rs
@@ -1,0 +1,16 @@
+use diesel::prelude::*;
+
+table! {
+    data_table (key) {
+        key -> Text,
+        value -> Text
+    }
+}
+
+table! {
+    metadata_table (key, blockhash) {
+        key -> Text,
+        blockhash -> Text,
+        value -> Text
+    }
+}
